+++ packages/aesthetic-adapter-aphrodite/src/NativeAdapter.js
@@ -21,6 +21,7 @@ export default class AphroditeAdapter extends Adapter {
 
     this.aphrodite = StyleSheet.extend([
       ...extensions,
+      { selectorHandler: this.handleHierarchySelector },
       { selectorHandler: this.handleGlobalSelector },
     ]);
   }
@@ -30,7 +31,37 @@ export default class AphroditeAdapter extends Adapter {
   }
 
   transform(...styles: StyleDeclaration[]): ClassName {
-    return this.aphrodite.css(...styles);
+    const legitStyles = [];
+    const tempStylesheet = {};
+    let counter = 0;
+
+    styles.forEach((style, i) => {
+      // eslint-disable-next-line no-underscore-dangle
+      if (style._name && style._definition) {
+        legitStyles.push(style);
+      } else {
+        tempStylesheet[`inline${counter}`] = style;
+        counter += 1;
+      }
+    });
+
+    if (counter > 0) {
+      legitStyles.push(...Object.values(this.create(tempStylesheet)));
+    }
+
+    return this.aphrodite.css(...legitStyles);
+  }
+
+  handleHierarchySelector(
+    selector: string,
+    baseSelector: string,
+    callback: (selector: string) => string | null,
+  ): string | null {
+    if (selector.charAt(0) !== '>' && selector.charAt(0) !== '[') {
+      return null;
+    }
+
+    return callback(`${baseSelector}${selector}`);
   }
 
   handleGlobalSelector(