+++ packages/aesthetic-adapter-glamor/CHANGELOG.md
@@ -1,4 +1,11 @@
-# 2.0.0
+# 2.1.0
+#### 🚀 New
+* Can now pass dynamic style objects to `transform` (will be akin to inline).
+* Unified Syntax
+  * Attribute selectors are now supported when starting with `[`.
+  * Direct descendant selectors are now supported with `>`.
+
+# 2.0.0 - 01/02/18
 #### 💥 Breaking
 * Rewritten to support the [core 2.0 changes](https://github.com/milesj/aesthetic/blob/master/packages/aesthetic/CHANGELOG.md).