+++ packages/aesthetic-adapter-typestyle/CHANGELOG.md
@@ -1,7 +1,17 @@
-# 1.0.1
+# 1.1.0
+#### 🚀 New
+* Can now pass dynamic style objects to `transform` (will be akin to inline).
+* Unified Syntax
+  * Attribute selectors are now supported when starting with `[`.
+  * Direct descendant selectors are now supported with `>`.
+
+#### 🐞 Fixed
+* Resolved some issues with pseudo classes.
+
+# 1.0.1 - 01/03/18
 #### 🛠 Internal
 * Tested against TypeStyle 1.6.
 
-# 1.0.0
+# 1.0.0 - 01/02/18
 #### 🎉 Release
 * Initial release!